Apply workspace Clippy policy and resolve strict lint findings
Inherit the existing workspace rules, remove redundant ownership conversions and clones, and address error sizes, lock scopes, and file open semantics. Preserve intentional asynchronous lock contention in storage regression tests.

Validated with cargo +1.94.1 clippy --workspace --all-targets --locked -- -D warnings and formatting checks. Focused storage, desktop frame queue, RDP helper, SSH configuration, and standalone russh buffer regressions passed.

+++ crates/oxideterm-gpui-terminal/src/app/interactions.rs
@@ -1167,7 +1167,7 @@ impl TerminalPane {
 
         let available = (geometry.track_height - geometry.height).max(px(1.0));
         let y = (position.y - geometry.y - thumb_offset_y).clamp(px(0.0), available);
-        let scroll_fraction = f32::from(y / available);
+        let scroll_fraction = y / available;
         let history = self.snapshot.scrollback_lines;
         let offset = ((1.0 - scroll_fraction) * history as f32).round() as usize;
         let snapshot = {
